Rename CollectionFolderGrid to CollectionFolderView
