+++ app/src/main/java/com/github/damontecres/wholphin/ui/cards/WatchedIcon.kt
@@ -10,6 +10,7 @@ import androidx.compose.material.icons.filled.Check
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.tv.material3.Icon
 import androidx.tv.material3.MaterialTheme
@@ -19,7 +20,10 @@ import com.github.damontecres.wholphin.ui.theme.LocalTheme
 import com.github.damontecres.wholphin.ui.theme.WholphinTheme
 
 @Composable
-fun WatchedIcon(modifier: Modifier = Modifier) {
+fun WatchedIcon(
+    modifier: Modifier = Modifier,
+    padding: Dp = 2.dp,
+) {
     Icon(
         imageVector = Icons.Default.Check,
         contentDescription = null,
@@ -28,7 +32,7 @@ fun WatchedIcon(modifier: Modifier = Modifier) {
             modifier
                 .background(WatchedIconBackground(), shape = CircleShape)
                 .border(.5.dp, Color.Black, CircleShape)
-                .padding(2.dp),
+                .padding(padding),
     )
 }
